Ab#29: Restructure starter package names
