+++ src/main/java/com/github/jferrater/opa/data/filter/spring/boot/starter/repository/hibernate/OpaGenericDataFilterDao.java
@@ -1,4 +1,4 @@
-package com.github.jferrater.opa.data.filter.spring.boot.starter;
+package com.github.jferrater.opa.data.filter.spring.boot.starter.repository.hibernate;
 
 import org.springframework.beans.factory.config.BeanDefinition;
 import org.springframework.context.annotation.Scope;
