+++ src/main/java/com/github/jferrater/opa/data/filter/spring/boot/starter/repository/jpa/AbstractOpaJpaDao.java
@@ -1,0 +1,4 @@
+package com.github.jferrater.opa.data.filter.spring.boot.starter.repository.jpa;
+
+public class AbstractOpaJpaDao {
+}
