+++ src/main/resources/META-INF/spring.factories
@@ -1,5 +1,5 @@
 org.springframework.boot.autoconfigure.EnableAutoConfiguration=\
         com.github.jferrater.opa.ast.db.query.config.OpaConfig,\
         com.github.jferrater.opa.ast.db.query.AppConfig, \
-        com.github.jferrater.opa.data.filter.spring.boot.starter.PersistenceConfig, \
-        com.github.jferrater.opa.data.filter.spring.boot.starter.OpaClientConfig
+        com.github.jferrater.opa.data.filter.spring.boot.starter.config.PersistenceConfig, \
+        com.github.jferrater.opa.data.filter.spring.boot.starter.config.OpaClientConfig
