+++ src/test/java/com/github/jferrater/opa/data/filter/spring/boot/starter/MyRepository.java
@@ -1,5 +1,6 @@
 package com.github.jferrater.opa.data.filter.spring.boot.starter;
 
+import com.github.jferrater.opa.data.filter.spring.boot.starter.repository.hibernate.OpaGenericDataFilterDao;
 import org.springframework.stereotype.Repository;
 import org.springframework.transaction.annotation.Transactional;
